% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\docType{methods}
\name{as.qspray}
\alias{as.qspray}
\alias{as.qspray,character-method}
\alias{as.qspray,qspray-method}
\alias{as.qspray,numeric-method}
\alias{as.qspray,bigz-method}
\alias{as.qspray,bigq-method}
\title{Coercion to a 'qspray' object}
\usage{
\S4method{as.qspray}{character}(x)

\S4method{as.qspray}{qspray}(x)

\S4method{as.qspray}{numeric}(x)

\S4method{as.qspray}{bigz}(x)

\S4method{as.qspray}{bigq}(x)
}
\arguments{
\item{x}{a \code{qspray} object or an object yielding a quoted integer or a 
quoted fraction after an application of \code{as.character}, e.g. a 
\code{bigq} number}
}
\value{
A \code{qspray} object.
}
\description{
Coercion to a 'qspray' object
}
\examples{
as.qspray(2)
as.qspray("1/3")
}
