% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\name{detQ}
\alias{detQ}
\title{Determinant of a rational matrix}
\usage{
detQ(M)
}
\arguments{
\item{M}{a square matrix such that each entry of \code{as.character(M)} is 
a quoted integer or a quoted fraction}
}
\value{
A quoted rational number representing the determinant.
}
\description{
Determinant of a square matrix with rational coefficients.
}
\examples{
M <- cbind(c("1/2", "3"), c("5/3", "-2/7"))
detQ(M)
}
