% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\name{ESFpoly}
\alias{ESFpoly}
\title{Elementary symmetric polynomial}
\usage{
ESFpoly(m, lambda)
}
\arguments{
\item{m}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
positive integers}
}
\value{
A \code{qspray} object.
}
\description{
Returns an elementary symmetric function as a polynomial.
}
\examples{
library(qspray)
ESFpoly(3, c(3, 1))
}
