% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{getConstantTerm}
\alias{getConstantTerm}
\alias{getConstantTerm,qspray-method}
\title{Get the constant term of a 'qspray' polynomial}
\usage{
\S4method{getConstantTerm}{qspray}(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A \code{bigq} number.
}
\description{
Get the constant term of a \code{qspray} polynomial.
}
