% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{involvedVariables}
\alias{involvedVariables}
\alias{involvedVariables,qspray-method}
\title{Variables involved in a 'qspray' polynomial}
\usage{
\S4method{involvedVariables}{qspray}(x)
}
\arguments{
\item{x}{a \code{qspray} object}
}
\value{
A vector of integers. Each integer represents the index of a 
  variable involved in \code{x}.
}
\description{
Variables involved in a \code{qspray} object.
}
\examples{
x <- qlone(1); z <- qlone(3)
involvedVariables(x^2 + x*z + 1) # should be c(1L, 3L)
}
