% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isConstant}
\alias{isConstant}
\alias{isConstant,qspray-method}
\title{Whether a 'qspray' polynomial is constant}
\usage{
\S4method{isConstant}{qspray}(x)
}
\arguments{
\item{x}{a \code{qspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{qspray} object defines a constant 
  polynomial.
}
