% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{isConstantQspray}
\alias{isConstantQspray}
\title{Whether a 'qspray' polynomial is constant}
\usage{
isConstantQspray(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{qspray} object defines a constant 
  polynomial.
}
