% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isQone}
\alias{isQone}
\alias{isQone,qspray-method}
\title{Whether a 'qspray' polynomial is the unit polynomial}
\usage{
\S4method{isQone}{qspray}(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{qspray} object defines the unit 
  polynomial.
}
