% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isQzero}
\alias{isQzero}
\alias{isQzero,qspray-method}
\title{Whether a 'qspray' polynomial is null}
\usage{
\S4method{isQzero}{qspray}(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{qspray} object defines the zero 
  polynomial.
}
