% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isUnivariate}
\alias{isUnivariate}
\alias{isUnivariate,qspray-method}
\title{Whether a 'qspray' is univariate}
\usage{
\S4method{isUnivariate}{qspray}(x)
}
\arguments{
\item{x}{a \code{qspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{qspray} object defines a
  univariate polynomial.
}
\note{
It is considered that a constant \code{qspray} is univariate, and 
  that the \code{qspray} object \code{y^2+1} where \code{y=qlone(2)} is 
  not univariate, although only one variable is present (see the note in 
  the documentation of \code{numberOfVariables}).
}
