% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{leadingCoefficient}
\alias{leadingCoefficient}
\title{Leading coefficient of a 'qspray' polynomial}
\usage{
leadingCoefficient(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
The coefficient of the leading term of \code{qspray},
  a \code{bigq} rational number.
}
\description{
Returns the leading coefficient of a \code{qspray} polynomial.
}
