% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{leadingTerm}
\alias{leadingTerm}
\title{Leading term of a 'qspray' polynomial}
\usage{
leadingTerm(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A list providing the exponents of the leading term in the field 
  \code{powers}, an integer vector, and the coefficient of the leading term 
  in the field \code{coeff}, a \code{bigq} rational number.
}
\description{
Returns the leading term of a \code{qspray} polynomial.
}
