% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\name{lone}
\alias{lone}
\title{Polynomial variable}
\usage{
lone(n)
}
\arguments{
\item{n}{nonnegative integer, the index of the variable}
}
\value{
A \code{qspray} object.
}
\description{
Create a polynomial variable.
}
\examples{
lone(2)
}
