% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\name{MSFpoly}
\alias{MSFpoly}
\title{Monomial symmetric function}
\usage{
MSFpoly(m, lambda)
}
\arguments{
\item{m}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
positive integers}
}
\value{
A \code{qspray} object.
}
\description{
Returns a monomial symmetric function as a polynomial.
}
\examples{
library(qspray)
MSFpoly(3, c(3, 1))
}
