% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{numberOfTerms}
\alias{numberOfTerms}
\alias{numberOfTerms,qspray-method}
\title{Number of terms in a 'qspray' polynomial}
\usage{
\S4method{numberOfTerms}{qspray}(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
An integer.
}
\description{
Number of terms of the polynomial defined by a 
  \code{qspray} object.
}
