% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{orderedQspray}
\alias{orderedQspray}
\title{Ordered 'qspray'}
\usage{
orderedQspray(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A \code{qspray} object. It defines the same polynomial as the 
  input \code{qspray} object but it is ordered.
}
\description{
Reorders the terms of a \code{qspray} object according to the 
  lexicographic order of the powers. This function is rather used 
  internally only but it is exported for internal usage in other packages.
}
\examples{
qspray <- rQspray()
qspray == orderedQspray(qspray) # should be TRUE
}
