% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\name{prettySymmetricQspray}
\alias{prettySymmetricQspray}
\title{Pretty symmetric qspray}
\usage{
prettySymmetricQspray(qspray, check = FALSE)
}
\arguments{
\item{qspray}{a \code{qspray} object, which should correspond to a 
symmetric polynomial}

\item{check}{Boolean, whether to check the symmetry}
}
\value{
A character string.
}
\description{
Prints a symmetric qspray polynomial as a linear combination of 
  the monomial symmetric polynomials.
}
\examples{
library(qspray)
qspray <- PSFpoly(4, c(3, 1)) + ESFpoly(4, c(2, 2)) + 4L
prettySymmetricQspray(qspray)
}
