% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\name{PSFpoly}
\alias{PSFpoly}
\title{Power sum polynomial}
\usage{
PSFpoly(m, lambda)
}
\arguments{
\item{m}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
positive integers}
}
\value{
A \code{qspray} object.
}
\description{
Returns a power sum function as a polynomial.
}
\examples{
library(qspray)
PSFpoly(3, c(3, 1))
}
