% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{qlone}
\alias{qlone}
\title{Polynomial variable}
\usage{
qlone(n)
}
\arguments{
\item{n}{positive integer, the index of the variable}
}
\value{
A \code{qspray} object.
}
\description{
Creates a polynomial variable. Using this function is the main 
  way to build \code{qspray} objects.
}
\examples{
x <- qlone(1)
y <- qlone(2)
(x + y) * (x - y)
}
