% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\name{qspray-unary}
\alias{qspray-unary}
\alias{+,qspray,missing-method}
\alias{-,qspray,missing-method}
\title{Unary operators for qspray objects}
\usage{
\S4method{+}{qspray,missing}(e1, e2)

\S4method{-}{qspray,missing}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{qspray}}

\item{e2}{nothing}
}
\value{
A \code{qspray} object.
}
\description{
Unary operators for qspray objects.
}
