% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{qspray_from_list}
\alias{qspray_from_list}
\title{(internal) Make a 'qspray' object from a list}
\usage{
qspray_from_list(qspray_as_list)
}
\arguments{
\item{qspray_as_list}{list returned by the Rcpp function 
\code{returnQspray}}
}
\value{
A \code{qspray} object.
}
\description{
This function is for internal usage. It is exported because 
  it is also used for internal usage in others packages.
}
