% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{rQspray}
\alias{rQspray}
\title{Random 'qspray'}
\usage{
rQspray()
}
\value{
A \code{qspray} object with at most 4 terms and at most 3 variables.
}
\description{
Generates a random \code{qspray} object.
}
