% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\name{SCHURcombination}
\alias{SCHURcombination}
\title{Symmetric polynomial in terms of the Schur polynomials}
\usage{
SCHURcombination(qspray, check = TRUE)
}
\arguments{
\item{qspray}{a \code{qspray} object defining a symmetric polynomial}

\item{check}{Boolean, whether to check the symmetry}
}
\value{
A list defining the combination. Each element of this list is a 
  list with two elements: \code{coeff}, a \code{bigq} number, and 
  \code{lambda}, an integer partition; then this list corresponds to the 
  term \code{coeff * SchurPol(n, lambda)}, where \code{n} is the number of 
  variables in the symmetric polynomial, and \code{SchurPol(n, lambda)} is 
  the Schur polynomial in \code{n} variables of \code{lambda}. This function 
  \code{SchurPol} is available in the \strong{jack} package.
}
\description{
Expression of a symmetric polynomial as a linear combination 
  of the Schur polynomials.
}
\examples{
qspray <- PSFpoly(4, c(3, 1)) + ESFpoly(4, c(2, 2)) + 4L
SCHURcombination(qspray)
}
