% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\docType{methods}
\name{showCoefficient}
\alias{showCoefficient}
\alias{showCoefficient,qspray-method}
\title{Function which prints the coefficients}
\usage{
\S4method{showCoefficient}{qspray}(x)
}
\arguments{
\item{x}{a \code{qspray} object}
}
\value{
A function which associates a character string to the coefficients.
}
\description{
This method is for internal usage. For any \code{qspray} 
  object, it returns a function which takes as argument a coefficient 
  and which returns the string formed by this coefficient surrounded by
  parentheses.
}
