% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showMonomial}
\alias{showMonomial}
\title{Print a monomial}
\usage{
showMonomial(x = "x")
}
\arguments{
\item{x}{a string, usually a letter such as \code{"x"} or \code{"X"}, to 
denote the variable}
}
\value{
A function which takes as argument a sequence of exponents and 
  which prints the corresponding monomial.
}
\description{
Print a monomial like \code{"x^(1, 0, 2)"}.
}
\examples{
showMonomial("X")(c(1, 0, 2))
}
