% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showMonomialOld}
\alias{showMonomialOld}
\title{Print a monomial}
\usage{
showMonomialOld(x = "x")
}
\arguments{
\item{x}{a string, usually a letter such as \code{"x"} or \code{"X"}, to 
denote the variable}
}
\value{
A function which takes as argument a sequence of exponents and 
  which prints the corresponding monomial.
}
\description{
Prints a monomial like \code{"x^(1, 0, 2)"}. This way of 
  showing a monomial was used by default in previous versions of this 
  package.
}
\examples{
showMonomialOld("X")(c(1, 0, 2))
showMonomialOld("X")(NULL)
}
\seealso{
\code{\link{showMonomialX1X2X3}}, \code{\link{showMonomialXYZ}}, 
  \code{\link{showQspray}}, \code{\link{showQsprayOption<-}}.
}
