% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showMonomialUnivariate}
\alias{showMonomialUnivariate}
\title{Print a univariate monomial}
\usage{
showMonomialUnivariate(x)
}
\arguments{
\item{x}{a string, usually a letter such as \code{"x"} or \code{"X"}, to 
denote the variable}
}
\value{
A function which takes as argument an exponent and 
  which prints the corresponding monomial.
}
\description{
Prints a univariate monomial.
}
\seealso{
\code{\link{showMonomialCanonical}}, 
  \code{\link{showQsprayUnivariate}}
}
