% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQsprayCanonical}
\alias{showQsprayCanonical}
\title{Print a 'qspray' object}
\usage{
showQsprayCanonical(var, ...)
}
\arguments{
\item{var}{a string, usually a letter such as \code{"x"} or \code{"X"}, 
which denotes the unindexed variables}

\item{...}{arguments passed to \code{\link{showQspray}}, such as 
\code{compact=TRUE}}
}
\value{
A function which prints a \code{qspray} object.
}
\description{
Print a \code{qspray} object given a string for the variable.
}
\examples{
qspray <- rQspray()
showQsprayCanonical("X")(qspray)
}
