% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQsprayUnivariate}
\alias{showQsprayUnivariate}
\title{Print a monomial for a univariate polynomial}
\usage{
showQsprayUnivariate(x)
}
\arguments{
\item{x}{a string, usually a letter such as \code{"x"} or \code{"X"}, to 
denote the variable}
}
\value{
A function which takes as argument an exponent and 
  which prints the corresponding monomial.
}
\description{
Prints a monomial like \code{"x^5"}.
}
