% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{composeQspray}
\alias{composeQspray}
\title{Compose 'qspray' polynomials}
\usage{
composeQspray(qspray, qsprays)
}
\arguments{
\item{qspray}{a \code{qspray} polynomial}

\item{qsprays}{a list containing \code{n} \code{qspray} polynomials where 
\code{n} is the number of variables of the polynomial given in the 
\code{qspray} argument}
}
\value{
The \code{qspray} polynomial obtained by composing the polynomial 
  given in the \code{qspray} argument with the polynomials given in the 
  \code{qsprays} argument.
}
\description{
Substitute the variables of a \code{qspray} polynomial with 
  some \code{qspray} polynomials.
}
\examples{
P <- qsprayMaker(string = "1/2 x^(1, 1) + 4 x^(0, 2)")
Q1 <- qsprayMaker(string = "x^(0, 1, 1)")
Q2 <- qsprayMaker(string = "2 x^(1) + x^(1, 1, 1)")
composeQspray(P, list(Q1, Q2))
}
