% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\name{derivQspray}
\alias{derivQspray}
\title{Partial derivative}
\usage{
derivQspray(qspray, i, derivative = 1)
}
\arguments{
\item{qspray}{object of class \code{qspray}}

\item{i}{dimension to differentiate with respect to}

\item{derivative}{how many times to differentiate}
}
\value{
A \code{qspray} object.
}
\description{
Partial derivative of a qspray polynomial.
}
\examples{
x
}
