% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\name{dQspray}
\alias{dQspray}
\title{Partial differentiation}
\usage{
dQspray(qspray, orders)
}
\arguments{
\item{qspray}{object of class \code{qspray}}

\item{orders}{integer vector, the orders of the differentiation}
}
\value{
A \code{qspray} object.
}
\description{
Partial differentiation of a qspray polynomial.
}
\examples{
library(qspray)
x <- qlone(1)
y <- qlone(2)
qspray <- x + 2*y  + 3*x*y
dQspray(qspray, c(1, 1))
derivQspray(derivQspray(qspray, 1), 2)
}
