% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\name{evalQspray}
\alias{evalQspray}
\title{Evaluate a 'qspray' object}
\usage{
evalQspray(qspray, values, exact = TRUE)
}
\arguments{
\item{qspray}{a \code{qspray} object}

\item{values}{vector of values; if \code{exact=TRUE}, each element of 
\code{as.character(values)} must be quoted integer or a quoted fraction, 
otherwise \code{values} must be a numeric vector}

\item{exact}{Boolean, whether to use \strong{gmp} arithmetic}
}
\value{
A \code{bigq} number if \code{exact=TRUE}, a double number otherwise.
}
\description{
Evaluation of the multivariate polynomial represented by a 
  \code{qspray} object.
}
\examples{
x <- lone(1); y <- lone(2)
P <- 2*x + "1/2"*y
evalQspray(P, c("2", "5/2", "99999")) # "99999" will be ignored
}
