% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getCoefficient}
\alias{getCoefficient}
\title{Get coefficient}
\usage{
getCoefficient(qspray, exponents)
}
\arguments{
\item{qspray}{a \code{qspray} object}

\item{exponents}{a vector of exponents}
}
\value{
The coefficient as a \code{bigq} number.
}
\description{
Get the coefficient corresponding to the given exponents.
}
\examples{
library(qspray)
x <- qlone(1)
y <- qlone(2)
p <- 2*x^2 + 3*y - 5
getCoefficient(p, 2)
getCoefficient(p, c(2, 0)) # same as getCoefficient(p, 2)
getCoefficient(p, c(0, 1))
getCoefficient(p, 0) # the constant term
getCoefficient(p, 3)
}
