% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getConstantTerm}
\alias{getConstantTerm}
\title{Get the constant term of a 'qspray' polynomial.}
\usage{
getConstantTerm(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A \code{bigq} number.
}
\description{
Get the constant term of a 'qspray' polynomial.
}
