% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{groebner}
\alias{groebner}
\title{Gröbner basis}
\usage{
groebner(G)
}
\arguments{
\item{G}{a list of qspray polynomials, the generators of the ideal}
}
\value{
A Gröbner basis of the ideal generated by \code{G}, given as a list 
  of qspray polynomials.
}
\description{
Returns a Gröbner basis following Buchberger's algorithm 
  using the lexicographical order.
}
\examples{
library(qspray)
f <- qsprayMaker(string = "x^(3) - 2 x^(1,1)")
g <- qsprayMaker(string = "x^(2,1) - 2 x^(0,2) + x^(1)")
groebner(list(f, g))
}
\references{
Michael Weiss, 2010. 
\href{Computing Gröbner Bases in Python with Buchberger’s Algorithm}{https://math.nyu.edu/degree/undergrad/ug_research/Weiss_SURE_Paper.pdf}.
}
