% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\name{HallInnerProduct}
\alias{HallInnerProduct}
\title{Hall inner product}
\usage{
HallInnerProduct(qspray1, qspray2, alpha = 1)
}
\arguments{
\item{qspray1, qspray2}{two symmetric \code{qspray} polynomials}

\item{alpha}{parameter equal to \code{1} for the usual Hall inner product, 
otherwise this is the "Jack parameter", an integer or a \code{bigq} number}
}
\value{
A \code{bigq} number.
}
\description{
Hall inner product of two symmetric polynomials.
}
