% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{isQone}
\alias{isQone}
\title{Whether a qspray polynomial is the unit polynomial}
\usage{
isQone(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{qspray} object defines the unit 
  polynomial.
}
