% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{isQzero}
\alias{isQzero}
\title{Whether a qspray polynomial is null}
\usage{
isQzero(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{qspray} object defines the zero 
  polynomial.
}
