% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{leadingTerm}
\alias{leadingTerm}
\title{Leading term of a 'qspray' polynomial}
\usage{
leadingTerm(qspray, d)
}
\arguments{
\item{qspray}{a \code{qspray} object}

\item{d}{integer, the number of variables of \code{qspray}, in general 
this is \code{numberOfVariables(qspray)}; setting a larger integer 
just adds some trailing zeros to the vector of exponents given in the 
output, and setting a smaller integer throws an error}
}
\value{
A list providing the exponents of the leading term in the field 
  \code{powers}, an integer vector, and the coefficient of the leading term 
  in the field  \code{coeff}, a \code{bigq} rational number.
}
\description{
Returns the leading term of a \code{qspray} polynomial.
}
