% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{numberOfTerms}
\alias{numberOfTerms}
\title{Number of terms in a 'qspray' polynomial}
\usage{
numberOfTerms(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
An integer.
}
\description{
Number of terms in the polynomial defined by a 
  \code{qspray} object.
}
