% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspray.R
\name{numberOfVariables}
\alias{numberOfVariables}
\title{Number of variables in a 'qspray' polynomial}
\usage{
numberOfVariables(qspray)
}
\arguments{
\item{qspray}{a \code{qspray} object}
}
\value{
An integer.
}
\description{
Number of variables involved in a \code{qspray} object.
}
