% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{numberOfVariables}
\alias{numberOfVariables}
\alias{numberOfVariables,qspray-method}
\title{Number of variables in a 'qspray' polynomial}
\usage{
\S4method{numberOfVariables}{qspray}(x)
}
\arguments{
\item{x}{a \code{qspray} object}
}
\value{
An integer.
}
\description{
Number of variables involved in a \code{qspray} object.
}
\note{
The number of variables in the \code{qspray} object \code{qlone(d)} 
  is \code{d}, not \code{1}.
}
