% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\name{PSPexpression}
\alias{PSPexpression}
\title{Symmetric polynomial in terms of the power sum polynomials.}
\usage{
PSPexpression(qspray)
}
\arguments{
\item{qspray}{a symmetric \code{qspray} polynomial (an error is returned if 
it is not symmetric)}
}
\value{
A \code{qspray} polynomial, say \eqn{P}, such that 
  \eqn{P(p_1, ..., p_n)} equals the input symmetric polynomial, 
  where \eqn{p_i} is the i-th power sum polynomial (\code{PSFpoly(n, i)}).
}
\description{
Expression of a symmetric \code{qspray} polynomial as a 
  polynomial in the power sum polynomials.
}
\note{
This function has been exported because it is used in the \strong{jack} 
  package.
}
