% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{qlone}
\alias{qlone}
\title{Polynomial variable}
\usage{
qlone(n)
}
\arguments{
\item{n}{nonnegative integer, the index of the variable}
}
\value{
A \code{qspray} object.
}
\description{
Create a polynomial variable.
}
\examples{
qlone(2)
}
