% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showMonomialX1X2X3}
\alias{showMonomialX1X2X3}
\title{Print a monomial}
\usage{
showMonomialX1X2X3(x)
}
\arguments{
\item{x}{a string, usually a letter such as \code{"x"} or \code{"X"}, to 
denote the unindexed variables}
}
\value{
A function which takes as argument a sequence of exponents and 
  which prints the corresponding monomial.
}
\description{
Prints a monomial like \code{"x1.x3^2"}.
}
\examples{
showMonomialX1X2X3("X")(c(1, 0, 2))
}
\seealso{
\code{\link{showQsprayX1X2X3}}, 
  \code{\link{showMonomialUnivariate}}, \code{\link{showQsprayOption<-}}.
}
