% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showMonomialXYZ}
\alias{showMonomialXYZ}
\title{Print a monomial}
\usage{
showMonomialXYZ(letters = c("x", "y", "z"), collapse = "")
}
\arguments{
\item{letters}{a vector of strings, usually some letters such as \code{"x"} 
and \code{"y"}, to denote the variables}

\item{collapse}{a string to denote the symbol representing the 
multiplication, e.g. \code{"*"} or \code{"."}}
}
\value{
A function which takes as argument a sequence of exponents and 
  which prints the corresponding monomial.
}
\description{
Prints a monomial like \code{"xz^2"}.
}
\examples{
showMonomialXYZ()(c(1, 0, 2))
showMonomialXYZ()(NULL)
showMonomialXYZ()(c(1, 0, 2, 3)) # no error
}
\seealso{
\code{\link{showQsprayXYZ}}, 
  \code{\link{showMonomialX1X2X3}}, \code{\link{showQsprayOption<-}}.
}
