% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQspray}
\alias{showQspray}
\title{Print a 'qspray' object}
\usage{
showQspray(showMonomial, compact = FALSE)
}
\arguments{
\item{showMonomial}{a function which takes as argument a sequence of 
exponents and which returns a string representing the corresponding 
monomial}

\item{compact}{a Boolean value}
}
\value{
A function which prints a \code{qspray} object.
}
\description{
Prints a \code{qspray} object given a function which prints 
  the monomials.
}
\seealso{
\code{\link{showQsprayCanonical}}, \code{\link{showMonomial}}, 
  \code{\link{showQsprayOption<-}}.
}
