% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQsprayCanonical}
\alias{showQsprayCanonical}
\title{Print a 'qspray' object}
\usage{
showQsprayCanonical(x, ...)
}
\arguments{
\item{x}{a string, usually a letter such as \code{"x"} or \code{"X"}, 
which denotes the unindexed variables}

\item{...}{arguments passed to \code{\link{showQspray}}, such as 
\code{compact=TRUE}}
}
\value{
A function which prints a \code{qspray} object.
}
\description{
Print a \code{qspray} object given a string for the variable.
}
\note{
The \code{show} method for \code{qspray} objects uses
  \code{showQsprayCanonical("x")} by default.
  But this can be controlled as follows. If a
  \code{qspray} object has an attribute \code{"x"}, then the value
  of this attribute will replace \code{"x"} in the \code{show} output.
}
\examples{
qspray <- rQspray()
showQsprayCanonical("X")(qspray)
}
