% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQsprayOption<-}
\alias{showQsprayOption<-}
\title{Set show option to a 'qspray' object}
\usage{
showQsprayOption(x, which) <- value
}
\arguments{
\item{x}{a \code{qspray} object}

\item{which}{which option to set; this can be \code{"x"}, 
\code{"showMonomial"}, or \code{"showQspray"}}

\item{value}{the value of the option}
}
\value{
This returns the updated \code{qspray}.
}
\description{
Set show option to a \code{qspray} object
}
\examples{
qspray <- rQspray()
showQsprayOption(qspray, "x") <- "a"
qspray
}
