% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQsprayUnivariate}
\alias{showQsprayUnivariate}
\title{Print a univariate polynomial}
\usage{
showQsprayUnivariate(x, ...)
}
\arguments{
\item{x}{a string, usually a letter such as \code{"x"} or \code{"X"}, to 
denote the variable}

\item{...}{arguments passed to \code{\link{showQspray}}, such as 
\code{compact=TRUE}}
}
\value{
A function which prints a univariate \code{qspray} object.
}
\description{
Prints a polynomial by printing monomials like \code{"x^5"}.
}
\seealso{
\code{\link{showMonomialUnivariate}}, \code{\link{showQspray}}, 
  \code{\link{showQsprayOption<-}}.
}
