% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQsprayX1X2X3}
\alias{showQsprayX1X2X3}
\title{Print a 'qspray' object}
\usage{
showQsprayX1X2X3(x = "x", collapse = ".", ...)
}
\arguments{
\item{x, collapse}{see \code{\link{showMonomialX1X2X3}}}

\item{...}{arguments passed to \code{\link{showQspray}}, such as 
\code{compact=TRUE}}
}
\value{
A function which prints a \code{qspray} object.
}
\description{
Prints a \code{qspray} object given a string for the variable.
}
\note{
The way to print \code{qspray} objects can be controlled with the 
 help of the function \code{\link{showQsprayOption<-}}.
}
\examples{
set.seed(3141)
( qspray <- rQspray() )
showQsprayX1X2X3("X")(qspray)
}
