% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showQsprayXYZ}
\alias{showQsprayXYZ}
\title{Print a polynomial}
\usage{
showQsprayXYZ(letters = c("x", "y", "z"), ...)
}
\arguments{
\item{letters}{a vector of strings, usually some letters such as \code{"x"}
and \code{"y"}, to denote the variables}

\item{...}{arguments passed to \code{\link{showQspray}}, such as 
\code{compact=TRUE}}
}
\value{
A function which prints a \code{qspray} object.
}
\description{
Prints a polynomial by printing monomials like \code{"x^2yz"}.
}
\examples{
set.seed(3141)
( qspray <- rQspray() )
showQsprayXYZ(c("X", "Y", "Z"))(qspray)

}
\seealso{
\code{\link{showMonomialXYZ}}, \code{\link{showQspray}}, 
  \code{\link{showQsprayOption<-}}.
}
